+++ app/frontend/package.json
@@ -28,11 +28,11 @@
   "dependencies": {
     "@bcgov/bc-sans": "^1.0.1",
     "@formio/vue": "github:formio/vue#v5.0.0-rc.2",
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/vue-fontawesome": "^3.0.3",
     "axios": "^1.4.0",
-    "bootstrap-scss": "^5.2.3",
+    "bootstrap-scss": "^5.3.1",
     "crypto-js": "^4.1.1",
     "fast-json-patch": "^3.1.1",
     "font-awesome": "^4.7.0",
