+++ app/frontend/src/router.js
@@ -1,12 +1,10 @@
 import NProgress from 'nprogress';
 import { createRouter, createWebHistory } from 'vue-router';
 
-import i18n from '~/internationalization';
-import { formService } from '~/services';
 import { useAuthStore } from '~/store/auth';
 import { useFormStore } from '~/store/form';
-import { useNotificationStore } from '~/store/notification';
-import { IdentityMode, IdentityProviders } from '~/utils/constants';
+import { IdentityProviders } from '~/utils/constants';
+import { preFlightAuth } from '~/utils/permissionUtils';
 
 let isFirstTransition = true;
 let router = undefined;
@@ -18,110 +16,6 @@ let router = undefined;
  * @returns {object} a Vue props object
  */
 const createProps = (route) => ({ ...route.query, ...route.params });
-
-/**
- * @function getErrorMessage
- * Gets the message to display for preflight errors. Expand this to add
- * friendlier messages for other errors.
- * @param {Object} options - The Object containing preflight request details.
- * @param {Error} error - The error that was produced.
- * @returns {string|undefined} - The error message to display, or undefined to
- *    use the default message.
- */
-function getErrorMessage(options, error) {
-  let errorMessage = undefined;
-  if (options.formId) {
-    const status = error?.response?.status;
-    if (status === 404 || status === 422) {
-      errorMessage = i18n.tc('trans.permissionUtils.formNotAvailable');
-    }
-  }
-  return errorMessage;
-}
-
-/**
- * @function preFlightAuth
- * Determines whether to enter a route based on user authentication state and idpHint
- * @param {Object} options Object containing either a formId or submissionId attribute
- * @param {Object} next The callback function
- */
-export async function preFlightAuth(options = {}, next) {
-  const notificationStore = useNotificationStore();
-  // Support lambda functions (Consider making them util functions?)
-  const getIdpHint = (values) => {
-    return Array.isArray(values) && values.length ? values[0] : undefined;
-  };
-  const isValidIdp = (value) =>
-    Object.values(IdentityProviders).includes(value);
-
-  // Determine current form or submission idpHint if available
-  let idpHint = undefined;
-  try {
-    if (options.formId) {
-      const { data } = await formService.readFormOptions(options.formId);
-      idpHint = getIdpHint(data.idpHints);
-    } else if (options.submissionId) {
-      const { data } = await formService.getSubmissionOptions(
-        options.submissionId
-      );
-      idpHint = getIdpHint(data.form.idpHints);
-    } else {
-      throw new Error(
-        i18n.tc('trans.permissionUtils.missingFormIdAndSubmssId')
-      );
-    }
-  } catch (error) {
-    // Halt user with error page, use alertNavigate for "friendly" messages.
-    const message = getErrorMessage(options, error);
-    if (message) {
-      // Don't display the 'An error has occurred...' popup notification.
-      notificationStore.alertNavigate(
-        'error',
-        i18n.tc('trans.permissionUtils.formNotAvailable')
-      );
-    } else {
-      notificationStore.addNotification({
-        text: i18n.tc('trans.permissionUtils.loadingFormErrMsg'),
-        consoleError: i18n.tc('trans.permissionUtils.loadingForm', {
-          options: options,
-          error: error,
-        }),
-      });
-      notificationStore.errorNavigate();
-    }
-
-    return; // Short circuit this function - no point executing further logic
-  }
-
-  const authStore = useAuthStore();
-
-  if (authStore.authenticated) {
-    const userIdp = authStore.identityProvider;
-
-    if (idpHint === IdentityMode.PUBLIC || !idpHint) {
-      next(); // Permit navigation if public or team form
-    } else if (isValidIdp(idpHint) && userIdp === idpHint) {
-      next(); // Permit navigation if idps match
-    } else {
-      const msg = i18n.t('trans.permissionUtils.idpHintMsg', {
-        idpHint: idpHint.toUpperCase(),
-      });
-      notificationStore.addNotification({
-        text: msg,
-        consoleError: msg,
-      });
-      notificationStore.errorNavigate(msg);
-    }
-  } else {
-    if (idpHint === IdentityMode.PUBLIC) {
-      next(); // Permit navigation if public form
-    } else if (isValidIdp(idpHint)) {
-      authStore.login(idpHint); // Force login flow with specified idpHint
-    } else {
-      authStore.login(); // Force login flow with user choice
-    }
-  }
-}
 
 /**
  * @function getRouter
